requirements.txt
XlsxWriter==1.3.7
Pillow==8.1.0